backend/requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
et_xmlfile==2.0.0
# Editable install with no version control (finance_ai==0.1)
-e d:\pyprojects\finance-ai-advisor\backend
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.2
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hupper==1.12.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
passlib==1.7.4
PasteDeploy==3.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pyramid==2.0.2
pyramid-mako==1.1.0
pyramid_debugtoolbar==4.12.1
pyramid_jinja2==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
threadpoolctl==3.6.0
tqdm==4.67.3
translationstring==1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
venusian==3.1.1
waitress==3.0.2
WebOb==1.8.9
xlrd==2.0.1
zope.deprecation==6.0
zope.interface==8.2
